:truck: Move resources
