+++ assets/html/ranking.html
@@ -17,7 +17,7 @@
       <div id="under_bar_content">        
         <bottom id="back_bth">←戻る</bottom>
         <div id="under_bar_content1">
-          <img src="../img/lay_mogura.png">
+          <img src="../assets/img/lay_mogura.png">
           <p id="under_bar_label">順位のTOP100までが表示されるよ！</p>
         </div>
       </div>
